Refactor retry policy implementation and namespaces
This commit removes references to the `UKHO.ADDS.EFS.Builder.S100.Infrastructure` namespace and introduces the `UKHO.ADDS.EFS.Domain.RetryPolicy` namespace across multiple files. Key changes include:

- Updated import statements in `CreateBatchNode.cs`, `DownloadFilesNode.cs`, and `ProductSearchNode.cs` to utilize the new retry policy approach.
- Adjustments in `Program.cs` to align with the new retry policy.
- Removal of the `Polly` package reference version `8.5.2` in `UKHO.ADDS.EFS.Builder.S100.csproj`, replaced with version `8.6.0` in `UKHO.ADDS.EFS.Domain.csproj`.
- Significant changes in `HttpClientPolicyLogger.cs` and `HttpClientPolicyProvider.cs`, including namespace updates and new using directives.
- Updates to `HttpClientPolicyProviderTest.cs` to reflect the new retry policy implementation.

These changes enhance the modularity and robustness of the retry policy mechanism, improving error handling and resilience in HTTP client operations.
