+++ src/UKHO.ADDS.EFS.Builder.S100/Program.cs
@@ -13,7 +13,7 @@ using UKHO.ADDS.EFS.Builder.S100.Services;
 using UKHO.ADDS.EFS.Configuration.Orchestrator;
 using UKHO.ADDS.EFS.Extensions;
 using UKHO.ADDS.Infrastructure.Pipelines.Nodes;
-using UKHO.ADDS.EFS.Builder.S100.Infrastructure;
+using UKHO.ADDS.EFS.Domain.RetryPolicy;
 using ILogger = Microsoft.Extensions.Logging.ILogger;
 
 namespace UKHO.ADDS.EFS.Builder.S100
